+++ tests/testthat/test-codebook_create.R
@@ -6,7 +6,7 @@ metadata <- metadata_create (
           ))
 
 names (metadata )
-test_codebook <- codebook_create (metadata )
+test_codebook <- codebook_create ( metadata )
 
 examples_dir <- system.file("examples", package = "retroharmonize")
 survey_list <- dir(examples_dir)[grepl("\\.rds", dir(examples_dir))]
@@ -15,12 +15,15 @@ example_surveys <- read_surveys(
   file.path( examples_dir, survey_list), 
   save_to_rds = FALSE)
 
+test_survey_codebook <- codebook_create  (survey = example_surveys[[1]])
+
 waves_codebook <- codebook_waves_create ( waves = example_surveys )
 
 
 test_that("correct codebook structure is returned", {
-  expect_true(all(names ( test_codebook )[! names ( test_codebook ) %in% names(metadata)] %in% c("entry","val_code_orig","val_label_orig","label_range")))
-  expect_true(all(names ( waves_codebook )[! names ( test_codebook ) %in% names(metadata)] %in% c("entry","val_code_orig","val_label_orig","label_range")))
+  expect_true(all(names (test_survey_codebook)[! names ( test_codebook ) %in% names(metadata)] %in% c("entry","val_code_orig","val_label_orig","label_range")))
+  expect_true(all(names (test_codebook)[! names ( test_codebook ) %in% names(metadata)] %in% c("entry","val_code_orig","val_label_orig","label_range")))
+  expect_true(all(names (waves_codebook)[! names ( test_codebook ) %in% names(metadata)] %in% c("entry","val_code_orig","val_label_orig","label_range")))
 })
 
 
@@ -37,6 +40,7 @@ test_that("correct codebook contents are returned", {
 
 test_that("error handling", {
   expect_error ( codebook_create( survey = data.frame()) )
+  expect_error ( codebook_create(metadata = NULL, survey = NULL ))
 })