+++ NEWS.md
@@ -1,5 +1,6 @@
 # retroharmoinze 0.1.18
 * New functions `codebook_create()` and `codebook_waves_create()`
+* Improved `retroharmonize_labelled_spss_survey()` `summary` method.
 * Much exception handling included in the import file (discovered problems with Caucasus Barometer files.)
 * Sometimes `haven::read_spss()` reads columns as labelled, even though they do not have a single label. In class `survey()` we convert them back to numeric or character types.